Allow multiple translation blocks per page
Before this, all translation blocks would build their own
translation model if there wasn't a translation object for that
locale already in the database. Meaning, that when adding a totally
new translation consisting of multiple blocks, the blocks would
each build their own object and save them to the database.

The CoffeeScript isn't pretty, might want to clean it up a bit
before opening a PR to upstream.